+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Changed minimum Python version to 3.10 and minimum Cantera version to 3.0.1
 
+### Removed
+
+- Removed the `convert` function and command line option, instead use Cantera's `yaml2ck` and `ck2yaml` 
+
 ### Fixed
 
 - Fixed bug in soln2ck.py where the reaction high rate was being used instead of the reaction low rate.